Fixing self shadow bug, removing LUT
Avatar was self shadowing because the shadow map reflected the avatar state from the previous frame.

The LUT for the shadow maps was a workaround to turn the x position into a color value. Removing the LUT in since we can just directly use that value in Metal instead of needing a texture go between. This also should slightly improve shadow quality over DX.

Additional LUT removal