Fixing memory pile up during intro movie
The intro movie keeps its own run loop which prevents our autorelease pool from collecting memory until the intro movie is done. This leads to a pile up of gigs (or tens of gigs) of memory as no framebuffer can be released.

Apple was nice enough to add NSAutoreleasePool to their list of bridged C++ classes though, so we can insert our own autorelease pool to capture that memory right into the pipeline.