Fixing performance with anisotropic texturing
Metal didn’t like the dynamic array lookup I was doing. Converting it it to static lookups. This has restored expected performance. Also converted the sample type to a function constant.

There is cleanup to do now. The fragment uniforms have been reduced to the alpha threshold, which is now awkwardly in a struct by itself. None of the layer data lives there anymore. In the pipeline, the encoding for shader lookup has gotten real messy.

I’m probably going to look at adding MSAA first, but the cleanup is really overdue. I’ve had to work through multiple issues already that happened because I tried to do minor tweaks to that code but forgot to touch a specific part.