Fixing device selection
This should align the framebuffer device and the render device on dual GPU Intel Macbook Pros. Previously it would try to draw to an Intel IGP framebuffer while rendering on the discrete card.

The way this is implemented is a bit of a hack. The Mac client app probably needs a deeper refactoring.