Fixing water on Nvidia/MSL/Fast Math
The water vertex shader needs to square as part of it’s reference math, but the value it needs to square can sometimes be negative. HLSL usually just does the “right” thing but MSL (and GLSL) need more direction and can treat squaring or square rooting a negative as undefined behavior. Absolute valuing first - in since we know that a square will always produce a positive.

Long term - this HLSL code was ported over directly. It’s possible that this code could be replaced by some other Metal functions like reflect. More research is required.