Fixes for problematic assets in Kadish
A cube map near or in the vault in Kadish has inconsistant cubic mipmap face sizes, and will cause Metal to crash. This change will cause Metal to manually compute mipmap sizes, and then load the data anyway. This is what DX does. The mipmaps affected don’t seem to be intended to be used.