Initial version of private Metal textures
Textures can now be stored in private storage in Metal - which means on discrete cards they can be VRAM only. This feature is disabled on Apple Silicon - but perhaps should be enabled. Even though Apple Silicon doesn’t have VRAM, the GPU might be able to store a private texture in a better storage format.