Fixes for WaveRip and vertex index streaming
Streaming of subsets of vertex indexes was broken, fixing. Because of triple buffering or the index buffers, there might still be bugs. A partial update of a buffer should be applied on the previous version of the buffer, but we have to buffer swap. Should consider copying the previous buffers contents into the new buffer before updating for consistancy.

WaveRip was also badly broken. Fixing.