WIP
More display helpers changes

plClient does querying of the possible display modes in it’s constructor. For now - using a singleton as a way to set a global display helper before this process starts.

Keeping display helpers in the plPipeline family. It’s possible it belongs at a lower level alongside hsG3DDeviceMode in since display helpers are platform and not render specific.

Properly subscribing to display change updates

WIP